+++ src/progressive_substitution.py
@@ -5,7 +5,7 @@ def progressive_substitution(ab):
     n = ab.shape[0]
     assert ab.shape[1] == n + 1
 
-    xs = np.zeros(n)
+    xs = np.zeros(n, dtype=np.float64)
     xs[0] = ab[0][n] / ab[0][0]
 
     for i in range(1, n):
